third_party/WebKit/LayoutTests/shadow-dom/v0/v0-inert-features-in-v1-shadow-trees.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>v0 features should be inert in v1 shadow trees.</title>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
<script src='../resources/shadow-dom.js'></script>
<div id='host'>
  <template data-mode='open'>
    <content></content>
    <shadow></shadow>
  </template>
  <div id='child'></div>
</div>
<script>
convertTemplatesToShadowRootsWithin(host);
test(function () {
    var content = host.shadowRoot.querySelector('content');
    var shadow = host.shadowRoot.querySelector('shadow');
    assert_equals(content.getDistributedNodes().length, 0);
    assert_equals(shadow.getDistributedNodes().length, 0);
}, "insertion points should be inert in v1");
</script>
